Section 11.120 - New input action for firing in the direction of the mouse
This is a divergent solution from the course content. The course had an input action for the modifier button and manually tracked whether the modifier button was pressed in the action handler and adjusted behavior accordingly. I opted against this solution, as it ties application code to specific input devices.

Instead, I have opted to create a new discrete input action/tag called "AttackTarget", which is triggered by a Chorded Input in the input context. This decouples the application code from the input method, and allows for future cases where the input method may not be a keyboard (for example, controller right stick + fire button).