[fix] Rebuild BP_POI_Fishing as a mostly native actor
Fishing has been broken since mid-July, when the AuraPOI actor was refactored. To prevent this sort of breakage again, I have migrated the Fishing POI to a mostly native implementation, with possible overrides in blueprint for each step in the fishing event. This resulting in the refactoring of a number of other components, including:

- FAuraFishCatch is now a UObject
-WBP_Fishing_Message is now backed by a native widget
- WBP_Interaction is now backed by a native widget
- BP_Ripple is now backed by a native actor
- IFishingComponentInterface has been eliminated

Note - this fix also includes migration of UPlayerInventoryComponent to PlayerState.