[feat] Move data out of GameMode, Leverage GameInstanceSubsystems for Data and Save Management (#82)
closes https://github.com/TimPoliquin/unreal-udemy-aura/issues/74

The introduction of save games was done at the cost of supporting
multiplayer in the Top Down RPG course. This large refactor hopes to
rectify that by moving all data and save game related functionality out
of GameMode and into GameInstanceSubsystems. Included are major
refactors to the Save Game system, level transition, abilities,
attributes, and the Fishing mini-game.

- Move SaveGame functionality out of GameMode into a new SaveGameManager
subsystem
- Move data out of GameMode into new AuraGameDataSubsystem
- Refactor character and world state loading - it is now triggered by
the SaveManager instead of by character possession
- Refactor FishingGame out of blueprint - it's been broken for months,
but was undetected
- Refactor Inventory so it is now globally shared across players
- Extract Equipment from Inventory so it can differ between players
- Refactor character progression so it is now globally shared across
players
- Migrate Load Screen to new Save Game Manager